Add filename support to yaml::Input.
The current yaml::Input constructor only takes a StringRef for its first
parameter: clients pass in raw string data. This means that, for the
typical use case of opening a file and parsing it, we throw away the
filename information and report errors as occurring in "YAML".

This commit adds an alterate yaml::Input constructor that takes a reference
to a MemoryBuffer, then extracts the buffer and whatever identifier is
associated with that buffer and passes them to the underlying yaml::Stream
as a MemoryBufferRef.